pyproject.toml
[tool.poetry]
name = "selenium-01"
version = "0.1.0"
description = ""
authors = ["居正中 <AlanJui.1960@gmail.com>"]
readme = "README.md"
packages = [{include = "selenium_01"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.1"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pylint = "^2.15.10"
selenium = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
